refine files name to queue
